requirements.txt
# Core OCR and PDF processing
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
regex>=2023.0.0

# VLM and AI dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0
huggingface-hub>=0.17.0

# Additional utilities for VLM processing
tqdm>=4.65.0
python-dotenv>=1.0.0
jsonschema>=4.17.0
typing-extensions>=4.5.0

# Optional: Flash Attention for better GPU performance
# flash-attn>=2.0.0  # Uncomment if using Flash Attention

# Optional: PDF to image conversion (alternative to PyMuPDF)
# pdf2image>=1.16.0

# Development tools
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
